moved integration test and removed sig checks
